Move the battery and uptime commands to a dedicated cog
Not currently working
