+++ src/bot.py
@@ -66,34 +66,6 @@ client.startTime = time.time()
 def admin(member):
     return True if member.id in client.admins else False
 
-async def checkBattery(client, limit):
-    flag = False
-    while True:
-        battery = psutil.sensors_battery()
-        if battery.percent < limit and not flag:
-            print('battery is at ', battery.percent)
-            flag = True
-            pings = ' '.join(str(client.get_user(user).mention) for user in client.admins)
-            await client.get_channel(899734389724942396).send(f'{pings} Battery is low! Please charge me!')
-        else:
-            flag = False
-        await asyncio.sleep(300)
-    return
-
-def convertSeconds(seconds):
-    days, seconds = divmod(seconds, 86400)
-    hours, seconds = divmod(seconds, 3600)
-    minutes, seconds = divmod(seconds, 60)
-    time_str = ""
-    if days > 0: time_str += f"{round(days)}d "
-    if hours > 0: time_str += f"{round(hours)}h "
-    if minutes > 0: time_str += f"{round(minutes)}m "
-    if seconds > 0: time_str += f"{round(seconds)}s"
-    return time_str
-
-def formatTime(timestamp):
-    return time.strftime("%d/%m/%Y %H:%M:%S", time.gmtime(timestamp + 8 * 3600))
-
 #endregion
 
 #region ----------------------------------------------------- EVENTS -------------------------------------------------
@@ -102,7 +74,6 @@ def formatTime(timestamp):
 async def on_ready():
     clear()
     print(f'{client.user} has connected to Discord!')
-    await client.get_user(285311305253126145).send(f'{client.user.mention} has connected to Discord!\n{formatTime(client.startTime)}')
     await checkBattery(client, 15)
     return
 
@@ -126,27 +97,6 @@ async def on_member_join(member):
 @client.slash_command(description='Will return "Pong" if the bot is online.')
 async def ping(interaction : nextcord.Interaction):
     await interaction.send(f'🏓 **Pong!** ({round(client.latency*1000)}ms)')
-    return
-
-@client.slash_command(description='Will return the battery of the bot.')
-async def battery(interaction : nextcord.Interaction):
-    battery = psutil.sensors_battery()
-    colour = nextcord.Colour.green() if battery.percent > 15 else nextcord.Colour.red()
-
-    embed = nextcord.Embed(title="Server Battery Stats", colour=colour)
-    embed.add_field(name='Battery percentage:', value=f'`{round(battery.percent, 2)}%`', inline=False)
-    embed.add_field(name='Power plugged in:', value=f'`{battery.power_plugged}`', inline=False)
-    embed.add_field(name='Battery time remaining:', value=f'`{convertSeconds(battery.secsleft)}`', inline=False)
-    await interaction.send(embed=embed)
-    return
-
-@client.slash_command(description='Will return the bot\'s Uptime')
-async def uptime(interaction : nextcord.Interaction):
-    uptime = convertSeconds(time.time() - client.startTime)
-    embed = nextcord.Embed(title=f"{client.user.name} Uptime ⌛", colour=nextcord.Colour.blue())
-    embed.add_field(name='Start Time:', value=f'`{formatTime(client.startTime)}`', inline=False)
-    embed.add_field(name='Uptime:', value=f'`{uptime}`', inline=False)
-    await interaction.send(embed=embed)
     return
 
 @client.slash_command(description='Help Command alias')
